db-backend/pyproject.toml
[project]
name = "db-backend"
version = "0.1.0"
description = "FastAPI backend for douban knowledge graph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.112.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "neo4j>=5.17.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "pymysql>=1.1.0",
    "dbutils>=3.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "bcrypt<5.0.0",
    "cachetools>=5.3.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.25.0",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
